# app/marksman/internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/level:
        post:
            tags:
                - Level
            operationId: Level_CreateLevel
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.CreateLevelRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.CreateLevelReply'
    /v1/level/{uid}:
        get:
            tags:
                - Level
            operationId: Level_GetLevel
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.LevelItem'
        put:
            tags:
                - Level
            operationId: Level_UpdateLevel
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateLevelRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateLevelReply'
        delete:
            tags:
                - Level
            operationId: Level_DeleteLevel
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.DeleteLevelReply'
    /v1/level/{uid}/status:
        put:
            tags:
                - Level
            operationId: Level_UpdateLevelStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateLevelStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateLevelStatusReply'
    /v1/levels:
        get:
            tags:
                - Level
            operationId: Level_ListLevel
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.ListLevelReply'
    /v1/levels/select:
        get:
            tags:
                - Level
            operationId: Level_SelectLevel
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SelectLevelReply'
...
    /v1/metric/strategy/{strategyUID}/level:
        post:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_SaveStrategyMetricLevel
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricLevelRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricLevelReply'
    /v1/metric/strategy/{strategyUID}/level/{levelUID}:
        get:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_GetStrategyMetricLevel
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: levelUID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.StrategyMetricLevelItem'
        delete:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_DeleteStrategyMetricLevel
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
                - name: levelUID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.DeleteStrategyMetricLevelReply'
...
components:
    schemas:
...
        marksman.api.v1.CreateLevelReply:
            type: object
            properties:
                uid:
                    type: string
        marksman.api.v1.CreateLevelRequest:
            type: object
            properties:
                name:
                    type: string
                remark:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                bgColor:
                    type: string
                type:
                    type: integer
                    format: enum
...
        marksman.api.v1.DeleteLevelReply:
            type: object
            properties: {}
...
        marksman.api.v1.DeleteStrategyMetricLevelReply:
            type: object
            properties: {}
...
        marksman.api.v1.LevelItem:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                status:
                    type: integer
                    format: enum
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                bgColor:
                    type: string
                type:
                    type: integer
                    format: enum
        marksman.api.v1.LevelItemSelect:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
...
        marksman.api.v1.ListLevelReply:
            type: object
            properties:
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.LevelItem'
...
        marksman.api.v1.SaveStrategyMetricLevelReply:
            type: object
            properties: {}
        marksman.api.v1.SaveStrategyMetricLevelRequest:
            type: object
            properties:
                strategyUID:
                    type: string
                levelUID:
                    type: string
                mode:
                    type: integer
                    format: enum
                condition:
                    type: integer
                    format: enum
                values:
                    type: array
                    items:
                        type: number
                        format: double
                duration:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                status:
                    type: integer
                    format: enum
...
        marksman.api.v1.SelectLevelReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.LevelItemSelect'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        marksman.api.v1.StrategyMetricLevelItem:
            type: object
            properties:
                levelUID:
                    type: string
                strategyUID:
                    type: string
                level:
                    $ref: '#/components/schemas/marksman.api.v1.LevelItem'
                mode:
                    type: integer
                    format: enum
                condition:
                    type: integer
                    format: enum
                values:
                    type: array
                    items:
                        type: number
                        format: double
                duration:
                    pattern: ^-?(?:0|[1-9][0-9]{0,11})(?:\.[0-9]{1,9})?s$
                    type: string
                status:
                    type: integer
                    format: enum
...
        marksman.api.v1.UpdateLevelReply:
            type: object
            properties: {}
        marksman.api.v1.UpdateLevelRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                remark:
                    type: string
                metadata:
                    type: object
                    additionalProperties:
                        type: string
                bgColor:
                    type: string
                type:
                    type: integer
                    format: enum
        marksman.api.v1.UpdateLevelStatusReply:
            type: object
            properties: {}
        marksman.api.v1.UpdateLevelStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
...